--- a/tsdb/querier.go
+++ b/tsdb/querier.go
@@ -269,12 +269,18 @@ func PostingsForMatchers(ix IndexReader, ms ...*labels.Matcher) (index.Postings,
 				if err != nil {
 					return nil, err
 				}
+				if index.IsEmptyPostings(it) {
+					return index.EmptyPostings(), nil
+				}
 				its = append(its, it)
 			} else { // l="a"
 				// Non-Not matcher, use normal postingsForMatcher.
 				it, err := postingsForMatcher(ix, m)
 				if err != nil {
 					return nil, err
+				}
+				if index.IsEmptyPostings(it) {
+					return index.EmptyPostings(), nil
 				}
 				its = append(its, it)
 			}